Sync 1.45 r1255: w105 mussrv
[01255 ]: w105 mussrv
[01254 ]: w105 linux make
https://sourceforge.net/p/doomlegacy/svn/1255/tree/

+++ SRC/DJGPPDOS/I_sound.c
@@ -448,7 +448,7 @@ void I_ShutdownMusic(void)
     music_started=false;
 }
 
-void I_PlaySong(int handle, int looping)
+void I_PlaySong(int handle, byte looping)
 {
     if(nomusic)
         return;
